refactor!: move security rule to separate pkg
Move the definition to a separate package, controller/nvapi/v1 and
change all the references.
